Add AlbumDurationCalculator
AlbumDurationCalculator and its test class were added to
be able to get the duration of a full album because
Spotify's api only provides durations of tracks. Also
the durations of tracks are only provided in milliseconds
so if the duration of an album was to be shown in hours
or minutes it would have to be converted to a duration is
returned by AlbumDurationCalculator's only method.

All of the classes were moved into "model" and "view"
packages for maintaining organization in the future.