Add ComboBox selection functionality
Another method was added to the input area listener so
that when the user chooses what information they want to
see in the combo box only the information view they
select is shown.

Most other GUI classes were refactored for clean code and
readability. The AlbumTrackListingFormatter was removed
because it was never used or needed.