Change SpotifyAlbumSearcher to get recommendations
SpotifyAlbumSearcher was renamed to SpotifyApiApplicant
to reflect the added ability for the object to be to
query spotify for ten recommendations based on an album.
Spotify's web API only allows us to search for
recommendations based on either a list of genres, an
artist's ID, or a track's ID. Instead of passing a
an ID to the method that gets recommendations, Album
objects are passed since the main idea of why we are
get recommendations from spotify was to get
recommendations for albums based on other albums. Also,
of an Album has multiple artists then passing the Album
object allows us to get the ID for all the artists and
potentially add genres or track IDs to the query
parameters of the recommendations query.