Add albums and getAlbums to AlbumRecommendations
The albums member variable was added so that
AlbumRecommendations objects would only have a .getAlbums
method because we don't care about the individual tracks
recommended by Spotify only the tracks' album. We had to
keep the tracks member variable in AlbumRecommendations
because I couldn't find a way for the JsonUtil inner
class to return AlbumSimplified objects rather than null.
I don't believe we have to worry about the tracks not
being initialized before albums because the builder has to
always be called first.